Move ToAudioStream functions to common class
