+++ src/VGAudio/Containers/Bxstm/BCFstmReader.cs
@@ -2,7 +2,6 @@
 using System.Linq;
 using System.Text;
 using VGAudio.Formats;
-using VGAudio.Formats.GcAdpcm;
 using VGAudio.Utilities;
 using static VGAudio.Utilities.Helpers;
 
@@ -58,69 +57,7 @@ namespace VGAudio.Containers.Bxstm
                 return structure;
             }
         }
-
-        public static IAudioFormat ToAudioStream(BxstmStructure structure)
-        {
-            switch (structure.Codec)
-            {
-                case BxstmCodec.Adpcm:
-                    return ToAdpcmStream(structure);
-                case BxstmCodec.Pcm16Bit:
-                    return ToPcm16Stream(structure);
-                case BxstmCodec.Pcm8Bit:
-                    return ToPcm8Stream(structure);
-            }
-            return null;
-        }
-
-        private static GcAdpcmFormat ToAdpcmStream(BxstmStructure structure)
-        {
-            var channels = new GcAdpcmChannel[structure.ChannelCount];
-
-            for (int c = 0; c < channels.Length; c++)
-            {
-                var channelBuilder = new GcAdpcmChannelBuilder(structure.AudioData[c], structure.Channels[c].Coefs, structure.SampleCount)
-                {
-                    Gain = structure.Channels[c].Gain,
-                    Hist1 = structure.Channels[c].Hist1,
-                    Hist2 = structure.Channels[c].Hist2
-                };
-
-                channelBuilder.WithLoop(structure.Looping, structure.LoopStart, structure.SampleCount)
-                    .WithLoopContext(structure.LoopStart, structure.Channels[c].LoopPredScale,
-                        structure.Channels[c].LoopHist1, structure.Channels[c].LoopHist2);
-
-                if (structure.SeekTable != null)
-                {
-                    channelBuilder.WithSeekTable(structure.SeekTable[c], structure.SamplesPerSeekTableEntry);
-                }
-
-                channels[c] = channelBuilder.Build();
-            }
-
-            return new GcAdpcmFormatBuilder(channels, structure.SampleRate)
-                .WithTracks(structure.Tracks)
-                .Loop(structure.Looping, structure.LoopStart, structure.SampleCount)
-                .Build();
-        }
-
-        private static Pcm16Format ToPcm16Stream(BxstmStructure structure)
-        {
-            short[][] channels = structure.AudioData.Select(x => x.ToShortArray(structure.Endianness)).ToArray();
-            return new Pcm16Format.Builder(channels, structure.SampleRate)
-                .WithTracks(structure.Tracks)
-                .Loop(structure.Looping, structure.LoopStart, structure.SampleCount)
-                .Build();
-        }
-
-        private static Pcm8SignedFormat ToPcm8Stream(BxstmStructure structure)
-        {
-            return new Pcm8Format.Builder(structure.AudioData, structure.SampleRate, true)
-                .WithTracks(structure.Tracks)
-                .Loop(structure.Looping, structure.LoopStart, structure.SampleCount)
-                .Build() as Pcm8SignedFormat;
-        }
-
+        
         private static void ReadHeader(BinaryReader reader, BCFstmStructure structure)
         {
             reader.Expect((ushort)0xfeff);
